% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartlett_oneway.R
\name{bartlett_oneway}
\alias{bartlett_oneway}
\title{Do oneway test setting the variation according to the Bartlett results}
\usage{
bartlett_oneway(x, y)
}
\arguments{
\item{x}{response variable}

\item{y}{grouping variable}
}
\value{
a table of statistical output
}
\description{
Do oneway test setting the variation according to the Bartlett results
}
