% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_below_detection_limit.R
\name{convert_below_detection_limit}
\alias{convert_below_detection_limit}
\title{Converts below detection limit
The function 'convert_below_detection_limit' converts the data points which are "<0,1", "<0,01", "<0,02", "<0,06" within the dataframe to their half values.}
\usage{
convert_below_detection_limit(data)
}
\arguments{
\item{data}{}
}
\value{
the dataframe with the converted datapoints
}
\description{
Converts below detection limit
The function 'convert_below_detection_limit' converts the data points which are "<0,1", "<0,01", "<0,02", "<0,06" within the dataframe to their half values.
}
