% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_CV.R
\name{CV}
\alias{CV}
\title{Coefficient of Variation (CV) calculation
Calculates by dividing the mean with the standard deviation}
\usage{
CV(x)
}
\arguments{
\item{x}{data vector from which the coefficient of variation (CV) will be calculated}
}
\value{
the coefficients of variation
}
\description{
Coefficient of Variation (CV) calculation
Calculates by dividing the mean with the standard deviation
}
