% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_bonferroni_to_adonis.R
\name{do_bonferroni_to_adonis}
\alias{do_bonferroni_to_adonis}
\title{Does bonferroni corrections on 3 pairs of permanova (adonis) results}
\usage{
do_bonferroni_to_adonis(x_nat, x_te, x_me)
}
\arguments{
\item{x_nat}{p value of the first pairwise comparison}

\item{x_te}{p value of the second pairwise comparison}

\item{x_me}{p value of the third pairwise comparison}
}
\value{
A table of results of the three pairwise comparisons Bonferroni corrected, with their p values, F values and degrees of Freedomi
}
\description{
Does bonferroni corrections on 3 pairs of permanova (adonis) results
}
