% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adonis_results.R
\name{get_adonis_results}
\alias{get_adonis_results}
\title{Gets the degress of freedom, p values and F values from the  Permanova computations.}
\usage{
get_adonis_results(x)
}
\arguments{
\item{x}{the adonis object to get the results from}
}
\value{
a data table that has the p values, f values and the degrees of freedom
}
\description{
Gets the degress of freedom, p values and F values from the  Permanova computations.
}
