% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_and_sd.R
\name{get_mean_and_sd}
\alias{get_mean_and_sd}
\title{Calculate mean and standard deviation of values with a grouping factor}
\usage{
get_mean_and_sd(data, response_variable, grouping_factor)
}
\arguments{
\item{data}{A data table}

\item{response_variable}{The response variable of which the mean+-sd calculation will be computed on}

\item{grouping_factor}{The grouping factor over which the mean+-sd calculation will be computed on}
}
\value{
A table of the results
}
\description{
Calculate mean and standard deviation of values with a grouping factor
}
