% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneway_test_results.R
\name{oneway_test_results}
\alias{oneway_test_results}
\title{Checks the equal variation and computes the oneway test accordingly}
\usage{
oneway_test_results(data, response_variable, grouping_factor, log_normalise)
}
\arguments{
\item{data}{the dataframe where the test will be conducted}

\item{response_variable}{the variable to be tested  in the format "response_variable"}

\item{grouping_factor}{the grouping variable  in the format "grouping_factor"}

\item{log_normalise}{logical, wether the response variable will be log scaled or not depending on the normal distribution of the data}
}
\value{
Gives the F values, p values and derees of freedom of the computed oneway test as well as the Barttlett test used to compute the variations
}
\description{
Checks the equal variation and computes the oneway test accordingly
}
