% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_pipeline_wrapper.R
\name{statistics_pipeline_wrapper}
\alias{statistics_pipeline_wrapper}
\title{Title}
\usage{
statistics_pipeline_wrapper(
  data,
  response_variable,
  grouping_factor,
  log_normalise
)
}
\arguments{
\item{data}{}

\item{response_variable}{}

\item{grouping_factor}{}

\item{log_normalise}{}
}
\value{
A table of adjusted p values
}
\description{
Title
}
