% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_results.R
\name{t_test_results}
\alias{t_test_results}
\title{F test followed by t.test according to the calculated equal/unequal variations}
\usage{
t_test_results(data, response_variable, grouping_factor, log_normalise)
}
\arguments{
\item{data}{A data.table where the var.test and t.test will be computed}

\item{response_variable}{the variable to be tested  in the format "response_variable"}

\item{grouping_factor}{the grouping variable  in the format "grouping_factor"}

\item{log_normalise}{logical, wether the response variable will be log scaled or not depending on the normal distribution of the data}
}
\value{
A table of F values, degrees of freedom and P values of the computes F test and t test results
}
\description{
F test followed by t.test according to the calculated equal/unequal variations
}
