% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_boxplot.R
\name{theme_boxplot}
\alias{theme_boxplot}
\title{Common theme function for the boxplots and pca plots}
\usage{
theme_boxplot()
}
\value{
ggplott theme
}
\description{
Common theme function for the boxplots and pca plots
}
