% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pca.R
\name{theme_pca}
\alias{theme_pca}
\title{Common theme function for the boxplots and pca plots}
\usage{
theme_pca()
}
\value{
ggplott theme
}
\description{
Common theme function for the boxplots and pca plots
}
