% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_test_t_test.R
\name{var_test_t_test}
\alias{var_test_t_test}
\title{Computes Two Sample t-test followed by F test 
F test compares variances which determines the var.equal statement in the following t-test}
\usage{
var_test_t_test(x, y)
}
\arguments{
\item{x}{response variable}

\item{y}{grouping variable}
}
\value{
A list of test statistics and mean values
}
\description{
Computes Two Sample t-test followed by F test 
F test compares variances which determines the var.equal statement in the following t-test
}
